Added a helper class for isometric geometry
The helper class just stores the angle of both axis on the isometric
plane as well as the length of each side of a grid square on that plane.

It can then return the grid position of a point on the x/y screen
plane.This requires the x/y posiiton of the bottom corner (the origin)
of the isometric plane and the x/y position itself.