Updated Isometric Logic
All the math in the last version was wrong. Got this one working so that
it can return the x/y coordinate of a point on the isometric plane. That
works. Just going to get it so that you can get the isometric coordinate
of an x/y coordinate. 

Also added some of the classes for representing the model. Just method
headers basically.