Got things drawing on the isometric axis
Fixed all the math again as it was wrong. Set up a helper method for
loading images for the items. It's just got one method for loading and
could go into the canvas class used for rendering.

Still figuring out the right coordinates to take for drawing. i.e. using
the isometric image to display.