Painters algorithm working for simple tiles
The painters algoirithm is implemented in the Window class. For it to
work you have to store everything in a grid poisiton of a 2d array. This
array is used for the rendering and nothing else. It dosn't worry about
the actual positions of thiings, you just say in which grid each object
is.