Changed painters algorithm, isoLogic and added some basic collision detection
Changed the painters algorithm to use the y coordinate of the item to be
drawn on the screen. This is much simpiler and does not require the use
of a 2D array. It now requires knowledge of their 'vertical' position.
Where things with higher vertical position (such as the moving square on
top of the tiles) is rendered first last (so it shows on top)

Modified the rounding being used in the isoLogic class to make it a bit
more accurate.

Added some basic collision detection between the moving item and the
walls. Just checks to see if their positions are equal.