Got rotation working, animation is abit weird though
Got the isoLogic class to support rotation of the axis. 

Animation however encounters issues when it comes to the painters
algorithm. The issue is that when moving the player during the animation
it is rendered before the tiles it is on and so appears under them.

The painters algorithm just uses values y coordinate to sort them. So it
shouldn't be an issue. Will have to look into it.